[CORDA-1778]: Moved ConfigSpecification to its right place.
